lib/python/pyproject.toml
[project]
name = "magic-import-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=10.4.0",
    "pdf2image>=1.17.0",
    "pymupdf>=1.24.9",
    "pypdf2>=3.0.1",
    "python-magic>=0.4.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
